refactor: split doctor.go monolith into domain-grouped files (#130)
